Avoid unlink call on empty string
Replaced move constructor by "swap" on default-constructed, empty
string. This because the std::string move constructor does not guarantee
that the moved string is actually erased. Also added test so we don't
unnecessarily call "unlink" on an empty string, even though this has no
effect.

+++ src/util/tempfile.cpp
@@ -164,5 +164,5 @@ Function: temporary_filet::~temporary_filet
 
 temporary_filet::~temporary_filet()
 {
-  unlink(name.c_str());
+  if(!name.empty()) unlink(name.c_str());
 }
